Move deviceufunc out of cuda directory
