+++ numba/cuda/tests/cudapy/test_gufunc_scheduling.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import, print_function, division
-from numba.cuda.deviceufunc import GUFuncEngine
+from numba.npyufunc.deviceufunc import GUFuncEngine
 from numba import unittest_support as unittest
